Workflow to check for missing keyword linking
