+++ scripts/python/src/fodt/keyword_uri_map_generator.py
@@ -159,12 +159,16 @@ def add_alias(kw_uri_map: dict[str, str], keyword: str, alias: str) -> None:
 #
 # SHELL USAGE:
 #
-# fodt-gen-kw-uri-map --maindir=<main_dir> --keyword_dir=<keyword_dir>
+#  fodt-gen-kw-uri-map --maindir=<main_dir> \
+#                      --keyword_dir=<keyword_dir> \
+#                      --check-changed
 #
 # DESCRIPTION:
 #
 #   Generates a map: KW_NAME -> URI for all keywords. The map is saved to the file
-#   "meta/kw_uri_map.txt" in the main directory.
+#   "meta/kw_uri_map.txt" in the main directory. If the --check-changed option is
+#   given, the script will only check if the files have changed and not write the
+#   output file. It will return a non-zero exit code if any files have changed.
 #
 # EXAMPLE:
 #
@@ -175,11 +179,20 @@ def add_alias(kw_uri_map: dict[str, str], keyword: str, alias: str) -> None:
 @click.command()
 @ClickOptions.maindir()
 @ClickOptions.keyword_dir
-def gen_kw_uri_map_cli(maindir: str|None, keyword_dir: str|None) -> None:
+@click.option('--check-changed', is_flag=True, help='Check if files have changed')
+def gen_kw_uri_map_cli(maindir: str|None, keyword_dir: str|None, check_changed: bool) -> None:
     logging.basicConfig(level=logging.INFO)
     keyword_dir = helpers.get_keyword_dir(keyword_dir)
     maindir = helpers.get_maindir(maindir)
     kw_uri_map = get_kw_uri_map(maindir, keyword_dir)
+    if check_changed:
+        orig_kw_uri_map = helpers.load_kw_uri_map(maindir)
+        if orig_kw_uri_map != kw_uri_map:
+            logging.error("Files have changed.")
+            exit(1)
+        else:
+            logging.info("Files have not changed.")
+            exit(0)
     with open(maindir / Directories.meta / FileNames.kw_uri_map, "w", encoding='utf8') as f:
         for kw in sorted(kw_uri_map.keys()):
             f.write(f"{kw} {kw_uri_map[kw]}\n")
